Make types of fromPlainModel() more strict in what they accept and make implementation more lenient
It is possible that fromPlainModel() is called with stuff returned from a web API. In that case, additional properties that were added in the API should not cause a runtime failure,
but we can just skip them. For APIs, usually addition of properties is not regarded as a breaking change, so we should not break on that either.